added sample plugin
the sample plugin defines the interface which should be implemented by
every plugin.

There surely is a problem with webpages where you have to click on a link
at the main page and then get the image embedded in the website.

This can be done with getNextUrl where it will autonoumously requests
content of that kind of links